Refactor: Remove old scorers and replace with optimised ones: Fix imports forall filters and scorers.
