+++ crawl4ai/__init__.py
@@ -46,8 +46,13 @@ from .deep_crawling import (
     DomainFilter,
     URLFilter,
     FilterStats,
-    FastKeywordRelevanceScorer,
-    FastURLScorer,
+    SEOFilter,
+    KeywordRelevanceScorer,
+    URLScorer,
+    CompositeScorer,
+    DomainAuthorityScorer,
+    FreshnessScorer,
+    PathDepthScorer,
     BestFirstCrawlingStrategy,
     DFSDeepCrawlStrategy,
     DeepCrawlDecorator,
@@ -64,8 +69,13 @@ __all__ = [
     "DomainFilter",
     "FilterStats",
     "URLFilter",
-    "FastKeywordRelevanceScorer",
-    "FastURLScorer",
+    "SEOFilter",
+    "KeywordRelevanceScorer",
+    "URLScorer",
+    "CompositeScorer",
+    "DomainAuthorityScorer",
+    "FreshnessScorer",
+    "PathDepthScorer",
     "DeepCrawlDecorator",
     "CrawlResult",
     "CrawlerHub",
