+++ tests/20241401/test_deep_crawl_filters.py
@@ -1,8 +1,5 @@
 from crawl4ai.deep_crawling.filters import ContentRelevanceFilter, URLPatternFilter, DomainFilter, ContentTypeFilter, SEOFilter
 async def test_pattern_filter():
-    import time
-    from itertools import chain
-
     # Test cases as list of tuples instead of dict for multiple patterns
     test_cases = [
         # Simple suffix patterns (*.html)
